api/requirements.txt
fastapi[standard]==0.117.1
python-dotenv==1.1.1

pymilvus==2.6.2
openai==1.108.1

pandas==2.3.2
PyPDF2==3.0.1
python-docx==1.2.0
python-pptx==1.0.2

uvicorn==0.36.0